fix(onboard): move smoke probe into inference probes
